Working on Changing Gatt in relation to Perspective
Tried making the cat and glow change to make the correct perspective using the distance between the index finger mcp and thumb mcp. But when implementing the code I didn’t take into account the fact that glow glow would squish and that when the hand would turn it would change the size. And when the hand moves closer to the camera it doesn’t work.Forgot to add the extra variable which made the glow turn into the correct circular chape, however it still isn’t centered. This is because the image made is set to corner so when I place it’s position on the middle finger mcp, it’s the corner of the png that is in the position, not the main image. But since the bg shadow uses corner mode perfectly, I’ve decided to change the position to the ring finger